requirements.txt
Django==3.1.1
django-allauth==0.41.0
django-crispy-forms==1.8.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
Pillow==6.2.1
django-cors-headers==3.2.0
django-rest-knox==4.1.0
python-social-auth==0.3.6
pusher_push_notifications==2.0.0
django-ckeditor==5.9.0
django-crispy-forms